apps/etl/requirements.txt
supabase-py==2.4.2
pandas==2.2.2
python-dotenv==1.0.1
pyyaml==6.0.2
supabase-py==2.4.2
yfinance==0.2.40
